delete discounts, started compose discounts

+++ code/src/main/java/domain/shop/predicate/ToBuildDiscountPredicate.java
@@ -23,7 +23,7 @@ public class ToBuildDiscountPredicate {
         predType = DiscountPredType.product;
     }
 
-    public ToBuildDiscountPredicate(int price) throws InvalidParamException {
+    public ToBuildDiscountPredicate(double price) throws InvalidParamException {
         if(amount < 0)
             throw new InvalidParamException("invalid amount");
         this.price = price;
